requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
langchain-text-splitters
langchain_huggingface
pandas
langchain-groq
langchain_core